refactor: Split out handler code
