+++ controlplane/src/bin/attune-server/api.rs
@@ -3,12 +3,9 @@ use std::{
     io::Write,
 };
 
-use aws_sdk_s3::config::BehaviorVersion;
 use axum::{
-    Json, Router,
-    extract::{DefaultBodyLimit, Multipart, Path, Query, State},
-    handler::Handler,
-    routing::{get, post},
+    Json,
+    extract::{Multipart, Path, Query, State},
 };
 use debian_packaging::deb::reader::{BinaryPackageEntry, BinaryPackageReader, ControlTarFile};
 use md5::Md5;
@@ -18,123 +15,32 @@ use sha2::{Digest, Sha256};
 use sqlx::types::{JsonValue, time::OffsetDateTime};
 use tabwriter::TabWriter;
 use time::format_description::well_known::Rfc2822;
-use tokio::signal;
-use tower_http::{trace::TraceLayer, validate_request::ValidateRequestHeaderLayer};
-use tracing::{Instrument, debug, debug_span, instrument};
-use tracing_subscriber::{
-    fmt::format::FmtSpan, layer::SubscriberExt as _, util::SubscriberInitExt as _,
-};
+use tracing::{Instrument, instrument, debug_span};
 
 #[derive(Clone, Debug)]
-struct ServerState {
-    db: sqlx::PgPool,
-    s3: aws_sdk_s3::Client,
+pub struct ServerState {
+    pub db: sqlx::PgPool,
+    pub s3: aws_sdk_s3::Client,
 
-    s3_bucket_name: String,
-    tenant_mode: TenantMode,
+    pub s3_bucket_name: String,
+    pub tenant_mode: TenantMode,
 }
 
 #[derive(Debug, Clone)]
-enum TenantMode {
+pub enum TenantMode {
     Single,
     Multi,
 }
 
-#[tokio::main]
-async fn main() {
-    // Initialize tracing.
-    tracing_subscriber::registry()
-        .with(
-            tracing_subscriber::fmt::layer()
-                .with_span_events(FmtSpan::NEW | FmtSpan::CLOSE)
-                .with_file(true)
-                .with_line_number(true)
-                .with_target(true)
-                .with_thread_ids(true)
-                .with_thread_names(true)
-                .with_writer(std::io::stderr)
-                .pretty(),
-        )
-        .with(tracing_subscriber::EnvFilter::from_default_env())
-        .init();
-
-    // Initialize upstream dependencies.
-    let db_url = std::env::var("ATTUNE_DATABASE_URL").expect("ATTUNE_DATABASE_URL not set");
-    let db = sqlx::postgres::PgPoolOptions::new()
-        .max_connections(5)
-        .connect(&db_url)
-        .await
-        .expect("could not connect to database");
-    let config = aws_config::defaults(BehaviorVersion::latest()).load().await;
-    let config = aws_sdk_s3::config::Builder::from(&config).build();
-    debug!(?config, "inferred AWS S3 configuration from environment");
-    let s3 = aws_sdk_s3::Client::from_conf(config);
-    let s3_bucket_name =
-        std::env::var("ATTUNE_S3_BUCKET_NAME").unwrap_or("attune-dev-0".to_string());
-
-    // Initialize configuration.
-    let secret = std::env::var("ATTUNE_SECRET").expect("ATTUNE_SECRET not set");
-    let tenant_mode =
-        std::env::var("ATTUNE_TENANT_MODE").map_or(TenantMode::Single, |v| match v.as_str() {
-            "single" => TenantMode::Single,
-            "multi" => TenantMode::Multi,
-            other => panic!("invalid ATTUNE_TENANT_MODE value: {}", other),
-        });
-
-    // Configure routes.
-    let api = Router::new()
-        .route(
-            "/repositories",
-            get(list_repositories).post(create_repository),
-        )
-        .route("/repositories/{repository_id}", get(repository_status))
-        .route(
-            "/repositories/{repository_id}/indexes",
-            get(generate_indexes),
-        )
-        .route("/repositories/{repository_id}/sync", post(sync_repository))
-        .route(
-            "/repositories/{repository_id}/packages",
-            get(list_packages)
-                .delete(remove_package)
-                .post(add_package.layer(DefaultBodyLimit::disable())),
-        );
-    let app = Router::new()
-        .nest("/api/v0", api)
-        .layer(ValidateRequestHeaderLayer::bearer(&secret))
-        .layer(TraceLayer::new_for_http())
-        .with_state(ServerState {
-            db,
-            s3,
-            s3_bucket_name,
-            tenant_mode,
-        });
-
-    // Start server.
-    println!("Listening on http://0.0.0.0:3000");
-    let listener = tokio::net::TcpListener::bind("0.0.0.0:3000").await.unwrap();
-    axum::serve(listener, app)
-        .with_graceful_shutdown(shutdown())
-        .await
-        .unwrap();
-}
-
-async fn shutdown() {
-    signal::unix::signal(signal::unix::SignalKind::terminate())
-        .expect("could not install SIGTERM handler")
-        .recv()
-        .await;
-}
-
 #[derive(Serialize)]
-struct Repository {
+pub struct Repository {
     id: i64,
     uri: String,
     distribution: String,
 }
 
 #[derive(Deserialize, Debug)]
-struct CreateRepositoryRequest {
+pub struct CreateRepositoryRequest {
     uri: String,
     distribution: String,
     origin: Option<String>,
@@ -147,7 +53,7 @@ struct CreateRepositoryRequest {
 
 #[axum::debug_handler]
 #[instrument(skip(state))]
-async fn create_repository(
+pub async fn create_repository(
     State(state): State<ServerState>,
     Json(payload): Json<CreateRepositoryRequest>,
 ) -> Json<Repository> {
@@ -195,7 +101,7 @@ async fn create_repository(
 
 #[axum::debug_handler]
 #[instrument(skip(state))]
-async fn list_repositories(State(state): State<ServerState>) -> Json<Vec<Repository>> {
+pub async fn list_repositories(State(state): State<ServerState>) -> Json<Vec<Repository>> {
     let repositories = sqlx::query!("SELECT id, uri, distribution FROM debian_repository")
         .fetch_all(&state.db)
         .await
@@ -213,12 +119,12 @@ async fn list_repositories(State(state): State<ServerState>) -> Json<Vec<Reposit
 }
 
 #[derive(Debug, Serialize)]
-struct RepositoryStatus {
+pub struct RepositoryStatus {
     changes: Vec<RepositoryChange>,
 }
 
 #[derive(Debug, Serialize)]
-struct RepositoryChange {
+pub struct RepositoryChange {
     package_id: i64,
     component: String,
     package: String,
@@ -231,7 +137,7 @@ struct RepositoryChange {
 
 #[axum::debug_handler]
 #[instrument(skip(state))]
-async fn repository_status(
+pub async fn repository_status(
     State(state): State<ServerState>,
     Path(repository_id): Path<u64>,
 ) -> Json<RepositoryStatus> {
@@ -271,13 +177,13 @@ async fn repository_status(
 }
 
 #[derive(Debug, Serialize)]
-struct RepositoryIndexes {
+pub struct RepositoryIndexes {
     release: String,
 }
 
 #[axum::debug_handler]
 #[instrument(skip(state))]
-async fn generate_indexes(
+pub async fn generate_indexes(
     State(state): State<ServerState>,
     Path(repository_id): Path<u64>,
 ) -> Json<RepositoryIndexes> {
@@ -593,14 +499,14 @@ async fn generate_indexes(
 }
 
 #[derive(Debug, Deserialize)]
-struct SyncRepositoryRequest {
+pub struct SyncRepositoryRequest {
     clearsigned: String,
     detached: String,
 }
 
 #[axum::debug_handler]
 #[instrument(skip(state))]
-async fn sync_repository(
+pub async fn sync_repository(
     State(state): State<ServerState>,
     Path(repository_id): Path<u64>,
     Json(payload): Json<SyncRepositoryRequest>,
@@ -794,7 +700,7 @@ async fn sync_repository(
 }
 
 #[derive(Deserialize, Serialize)]
-struct Package {
+pub struct Package {
     id: i64,
     package: String,
     version: String,
@@ -803,7 +709,7 @@ struct Package {
 
 #[axum::debug_handler]
 #[instrument(skip(state, multipart))]
-async fn add_package(
+pub async fn add_package(
     State(state): State<ServerState>,
     Path(repository_id): Path<u64>,
     Query(params): Query<HashMap<String, String>>,
@@ -1047,12 +953,12 @@ async fn add_package(
 
 #[axum::debug_handler]
 #[instrument]
-async fn list_packages() -> Json<Vec<Package>> {
+pub async fn list_packages() -> Json<Vec<Package>> {
     todo!()
 }
 
 #[axum::debug_handler]
 #[instrument]
-async fn remove_package() -> Json<Package> {
+pub async fn remove_package() -> Json<Package> {
     todo!()
 }
